Move Dnd Provider to Grid component
