+++ src/app/[locale]/dashboard/@board/Board.tsx
@@ -10,16 +10,7 @@ import PrintIcon from '@mui/icons-material/Print';
 import BookmarkBorderIcon from '@mui/icons-material/BookmarkBorder';
 import Grid from './Grid';
 import board from './Grid/testBoard.json';
-
-import { DndProvider } from 'react-dnd';
-import { TouchBackend } from 'react-dnd-touch-backend';
 import { ReactNode } from 'react';
-
-const dndOptions = {
-  enableTouchEvents: true,
-  enableMouseEvents: true,
-  enableKeyboardEvents: true,
-};
 
 export default function Board() {
   const renderTileFixedBoard = (item: {
@@ -117,18 +108,16 @@ export default function Board() {
           }}
         >
           <Box sx={{ maxWidth: '1500px', overflowX: 'scroll', height: '100%' }}>
-            <DndProvider backend={TouchBackend} options={dndOptions}>
-              <Grid
-                order={board.grid ? board.grid.order : []}
-                items={board.tiles}
-                columns={board.grid ? board.grid.columns : 6} //DEFAULT_COLUMNS_NUMBER}
-                rows={board.grid ? board.grid.rows : 6} // DEFAULT_ROWS_NUMBER}
-                dragAndDropEnabled={true} //{isSelecting}
-                renderItem={(item) => renderTileFixedBoard(item)}
-                onItemDrop={onTileDrop}
-                renderEmptyCell={renderEmptyCell}
-              />
-            </DndProvider>
+            <Grid
+              order={board.grid ? board.grid.order : []}
+              items={board.tiles}
+              columns={board.grid ? board.grid.columns : 6} //DEFAULT_COLUMNS_NUMBER}
+              rows={board.grid ? board.grid.rows : 6} // DEFAULT_ROWS_NUMBER}
+              dragAndDropEnabled={true} //{isSelecting}
+              renderItem={(item) => renderTileFixedBoard(item)}
+              onItemDrop={onTileDrop}
+              renderEmptyCell={renderEmptyCell}
+            />
           </Box>
         </Box>
       </Box>
